# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "@openapi:info",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/macros": {
      "get": {
        "description": "List all macros in reef-pi.",
        "tags": [
          "Macros"
        ],
        "summary": "List all macros.",
        "operationId": "macroList",
        "responses": {
          "200": {
            "description": "macro",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/macro"
              }
            }
          }
        }
      },
      "put": {
        "description": "Create a new macro.",
        "tags": [
          "Macros"
        ],
        "summary": "Create a macro.",
        "operationId": "macroCreate",
        "parameters": [
          {
            "description": "The macro to create",
            "name": "macro",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/macro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/macros/{id}": {
      "get": {
        "description": "Get an existing macro by id.",
        "tags": [
          "Macros"
        ],
        "summary": "Get a macro by id.",
        "operationId": "macroGet",
        "parameters": [
          {
            "description": "The Id of the macro",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/macro"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "Update an existing macro.",
        "tags": [
          "Macros"
        ],
        "summary": "Update a macro.",
        "operationId": "macroUpdate",
        "parameters": [
          {
            "description": "The Id of the macro to update",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The macro to update",
            "name": "macro",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/macro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "Delete an existing macro.",
        "tags": [
          "Macros"
        ],
        "summary": "Delete a macro.",
        "operationId": "macroDelete",
        "parameters": [
          {
            "description": "The Id of the macro to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/macros/{id}/revert": {
      "post": {
        "description": "Revert a macro in reverse.",
        "tags": [
          "Macros"
        ],
        "summary": "Reverse a macro.",
        "operationId": "macroRevert",
        "parameters": [
          {
            "description": "The Id of the macro to revert",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/macros/{id}/run": {
      "post": {
        "description": "Run a macro.",
        "tags": [
          "Macros"
        ],
        "summary": "Run a macro.",
        "operationId": "macroRun",
        "parameters": [
          {
            "description": "The Id of the macro to run",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Step": {
      "type": "object",
      "properties": {
        "config": {
          "type": "object",
          "x-go-name": "Config"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/reef-pi/reef-pi/controller/modules/macro"
    },
...
    "macro": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "reversible": {
          "type": "boolean",
          "x-go-name": "Reversible"
        },
        "steps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Step"
          },
          "x-go-name": "Steps"
        }
      },
      "x-go-name": "Macro",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/modules/macro"
    },
...
  }
}